This bit in the README may send people to Apigee in error.
I think so because I inquired with Apigee and learned the following:

Increased rate limits for consumers of the Twitter API via Apigee were
previously offered with the Apigee FREEMIUM service. This FREE traffic
is now being routed to a different set of Twitter APIs that do not allow
that increase.

AS a side note, It is not part and never has been available with the
Apigee Enterprise product.