Fix a few documentation errors
There were a few places in the friends_and_followers.rb documentation where the same text - "Returns an array of numeric IDs for every user the authenticated user is following" - must have been copy/pasted from another method, but it actually is incorrect for the method it was listed under.  For example, the 'friends' method does not return only the IDs, but actually returns the full collection of user objects.