Explicitly state Site Streaming API has limited access
Site Streaming API access is limited. This is not immediately
clear in the documentation/comments of the gem, although it
is clear in the twitter API. Adding a @note similar to the firehose
streaming API will help clear up any confusion that the
developer is misusing the API through the gem. Also added
restriction information to README for clarification.