Chomp empty URL fragments from the end of websites
This fixes a nasty bug that occurred when a URL fragment in a Twitter
user's website is present but empty (e.g. "http://example.com#"). In
this case, the @normalized_fragment instance variable in an
Addressable::URI object is set every time the normalized_fragment method
is invoked, setting the instance variable, thus violating the object's
frozen state.

Closes https://github.com/sferik/t/issues/197.