Add Website Cards functionality.
I coded this off of the docs but they're pretty
weird/inconsistent/wrong? Additional work will need to be done to ensure
things are properly stable.

Some things that should maybe be params were shifted to options in the
interest of not having functions with 5 args preferring the functions be
usable rather than catch missing args via arity checks.