Create Twitter::Ads::Cursor.
The cursor on the ads api behaves just slightly differently from the one
in the standard REST api. I've subclassed it and overridden the default
cursor value and next cursor methods to be compatible.

A straight-forward example usage has been added to the campaigns index
endpoint. The rest will follow.