Add `x-request-cost` to Twitter::RateLimit
Also expose `#rate_limit` method on Twitter::Stats. I'm not really happy
with this method of accomplishing this but it seems good enough for now.
It seems fairly important to have the rate limit information available
while making requests to the twitter API so I'm sticking the
Twitter::RateLimit object for the request that generated it into each
created Twitter::Stats. Perhaps further consideration will indicate that
it's not necessary until the rate limit is reached.