Adds development environment to configuration objects
With the new subscription APIs, including the Premium Search API, developers
need to configure a dev environment for the endpoint. The label for this
environment becomes part of the URL for the endpoint, like this:

`https://api.twitter.com/1.1/tweets/search/30day/my_env_name.json`

Adding the dev environment to the config makes it available whenever needed.