Use heuristic_parse for untrusted URLs.
Addressable::URI.parse will throw exceptions for URLs it thinks are
invalid. The issue with this is that Addressable and Twitter do not
agree on what qualifies as a valid URL. So a tweet can contain URL
entities that Addressable believes are invalid.
Addressable::URI.heuristic_parse is Addressable's more lenient parser.

This will make it so any tainted user data is parsed with
heuristic_parse. This way there is less of a chance of encountering an
Addressable::URI::InvalidURIError exception in the wild.